feat(aggregate/interactor): idempotency methods abstraction and segregation
